Some changes in the level1 and images to level 3

+++ lib/widgets/functional/level-1/car.dart
@@ -6,7 +6,7 @@ import 'package:flutter/cupertino.dart';
 
 class Car extends SpriteComponent {
   Size dimenstions;
-  int postion;
+  double postion;
   int ypostion;
   double maxY;
